requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# visual-spatial-reasoning/feature_extraction/lxmert/demo.ipynb: 3
# visual-spatial-reasoning/feature_extraction/lxmert/extract_img_features.py: 3
# visual-spatial-reasoning/feature_extraction/lxmert/processing_image.py: 23
# visual-spatial-reasoning/feature_extraction/lxmert/utils.py: 38
# visual-spatial-reasoning/src/data.py: 7
Pillow == 9.1.0

# visual-spatial-reasoning/feature_extraction/lxmert/utils.py: 45
PyYAML == 6.0

# visual-spatial-reasoning/feature_extraction/lxmert/extracting_data.py: 9
datasets == 2.1.0

# visual-spatial-reasoning/feature_extraction/visualbert/utils.py: 12,13,14,15,16,18,19,20,21
detectron2 == 0.5+cu111

# visual-spatial-reasoning/feature_extraction/lxmert/utils.py: 44
filelock == 3.6.0

# visual-spatial-reasoning/feature_extraction/lxmert/demo.ipynb: 2
# visual-spatial-reasoning/feature_extraction/lxmert/extract_img_features.py: 2
ipython == 8.3.0

# visual-spatial-reasoning/feature_extraction/lxmert/visualizing_image.py: 21,22,23,26
# visual-spatial-reasoning/feature_extraction/visualbert/utils.py: 5
# visual-spatial-reasoning/src/data.py: 9
# visual-spatial-reasoning/src/test.py: 5
# visual-spatial-reasoning/src/train.py: 8
matplotlib == 3.5.1

# visual-spatial-reasoning/feature_extraction/lxmert/demo.ipynb: 7
# visual-spatial-reasoning/feature_extraction/lxmert/extract_img_features.py: 9
# visual-spatial-reasoning/feature_extraction/lxmert/extracting_data.py: 10
# visual-spatial-reasoning/feature_extraction/lxmert/modeling_frcnn.py: 25
# visual-spatial-reasoning/feature_extraction/lxmert/processing_image.py: 21
# visual-spatial-reasoning/feature_extraction/lxmert/utils.py: 37
# visual-spatial-reasoning/feature_extraction/lxmert/visualizing_image.py: 24
# visual-spatial-reasoning/feature_extraction/visualbert/extract_img_features.py: 3
# visual-spatial-reasoning/feature_extraction/visualbert/utils.py: 8
# visual-spatial-reasoning/src/test.py: 8
# visual-spatial-reasoning/src/train.py: 6
numpy == 1.22.3

# visual-spatial-reasoning/feature_extraction/lxmert/utils.py: 41
# visual-spatial-reasoning/feature_extraction/lxmert/visualizing_image.py: 28
# visual-spatial-reasoning/feature_extraction/visualbert/utils.py: 7
# visual-spatial-reasoning/src/data.py: 3
# visual-spatial-reasoning/src/test.py: 7
# visual-spatial-reasoning/src/train.py: 2
opencv_python == 4.5.5.62

# visual-spatial-reasoning/data/splits/create_zeroshot_splits.py: 5
pandas == 1.4.2

# visual-spatial-reasoning/feature_extraction/lxmert/utils.py: 42
requests == 2.27.1

# visual-spatial-reasoning/feature_extraction/lxmert/demo.ipynb: 6
# visual-spatial-reasoning/feature_extraction/lxmert/extract_img_features.py: 7
# visual-spatial-reasoning/feature_extraction/lxmert/extracting_data.py: 11
# visual-spatial-reasoning/feature_extraction/lxmert/modeling_frcnn.py: 26,27,28
# visual-spatial-reasoning/feature_extraction/lxmert/processing_image.py: 22,24
# visual-spatial-reasoning/feature_extraction/lxmert/utils.py: 49,57
# visual-spatial-reasoning/feature_extraction/lxmert/visualizing_image.py: 25
# visual-spatial-reasoning/feature_extraction/visualbert/utils.py: 4,10
# visual-spatial-reasoning/src/data.py: 10,11
# visual-spatial-reasoning/src/eval.py: 5
# visual-spatial-reasoning/src/lxmert_for_classification.py: 3
# visual-spatial-reasoning/src/test.py: 4
# visual-spatial-reasoning/src/train.py: 13
torch == 1.11.0+cu113

# visual-spatial-reasoning/feature_extraction/lxmert/modeling_frcnn.py: 30
# visual-spatial-reasoning/feature_extraction/visualbert/utils.py: 4
# visual-spatial-reasoning/src/test.py: 4
torchvision == 0.12.0+cu113

# visual-spatial-reasoning/feature_extraction/lxmert/extract_img_features.py: 10
# visual-spatial-reasoning/feature_extraction/lxmert/utils.py: 39
# visual-spatial-reasoning/feature_extraction/visualbert/extract_img_features.py: 4
# visual-spatial-reasoning/src/data.py: 6
# visual-spatial-reasoning/src/eval.py: 4
# visual-spatial-reasoning/src/train.py: 9
tqdm == 4.63.0

# visual-spatial-reasoning/feature_extraction/lxmert/demo.ipynb: 13
# visual-spatial-reasoning/src/eval.py: 93
# visual-spatial-reasoning/src/lxmert_for_classification.py: 4
# visual-spatial-reasoning/src/test.py: 3
# visual-spatial-reasoning/src/train.py: 171
transformers == 4.17.0

# visual-spatial-reasoning/src/test.py: 2
utils == 1.0.1

# visual-spatial-reasoning/src/train.py: 4
wandb == 0.12.11

# visual-spatial-reasoning/feature_extraction/lxmert/demo.ipynb: 14
# visual-spatial-reasoning/feature_extraction/lxmert/extract_img_features.py: 16
# visual-spatial-reasoning/feature_extraction/lxmert/utils.py: 43
wget == 3.2
